Commit: Update Readme file
Explain:
I added a few things to the project readme file to specify certain
things for the project including overall idea behind the project,
different features that are going to be in the project, and a basic
runthough of the entire project. This file can server as a good
foundation for knowing what to build and how to move along when building
the project.