Commit: added design files
Explain:
I added a few documents and folders got the project initial designs.
these files include a general run though of the add to understand the
end image of how the app works, the features that include all of the
different things that are going to come with this project, and the
general idea of what the app is going to be based on and come back to.
Also included is the general idea for the first database schema that I
came up with in regards to the end schema fo the Alpha release.