Commit: Added comments for inital models
Explain:
for several of the columns in the inital model design for the database
schema, there are a set number of options that are predetermined and set
accountly. I added comments under the columns that have the set options
and what each of the values is to waht is saved in the database. I also
wanted to clarify what information is set by the project server and what
is inputted and set by the user himself. this will make it easier for
the develops to know what is coming and going as well as the source of
informaiton.