Commit: fixed some errors and added files
Explain:
The first change was done in the settings file. I had to add some changes
to the middleware portion becasue tehre was an error that was occuring
with the modules. The sencond major change was the change in the forms
file. I changed the Create group form as well as added the form for adding
users at each level of the users count. this was so that based on what the
user selects, displays a specific form.