Commit: New functionality
Explain:
I have now added a working version of the group creation function for the
project. You can now go the localhost/tab/createGroup link and fill out
the form which will create the group and save the record. You will then be
redirected to a page that has a list of hte groups and users with all of
the database informaiton that related to each of the grouips. It also
create a member record for the person who created the group(done by ther
server). I will be adding comments for all of the code that i have added
to the project for all developer and future developers to understand what
is going on with the code that i wrote. As of right now the function of
creating a groups is working. The next thing is adding comments and
starting on the adding members to the group.