Commit: new functionality
Explain:
I have now added a basic working function that will allow a user to add
new members to a specific group. As of right now you have to add one
member at a time until i add the javascript to the html that allows me to
auto add new input fields dynamically. I will be adding new comments and
basic struture notes within the code and push it again to make sure that
other developers have the ability to know what i am writing and why i am
writing it.

There was a small issue when it came to the reverse httpresponseredirect
and i was not able to really figure out the issue, so to add a memeber to
the gorup or access that view, you will have to physically add the group
name to the url and it will be handled properly once that is done: sample
url would look like this: ;127.0.0.1:8000/tab/groupName/addMembers. Every
time the form is submitted, you will be able to see  all the group sand
members to verify that it is working properly.