Commit: Added a users home page for development
Explain:
I added a new view that will display a uers file that will display all of
the different groups that the person is member of. For development this
allows me to click and go to the different group pages that he or she is
a member of and makes the groups details easily accessable. It was getting
complicated to get and develop for the groups home page and the members to
further develop the adding records functionality. Each of the group0s that
are shown up on the users home will be clickable and go to the groups
different pages.