Commit: added a group home page
Explain:
I created a gorup home page for all of the different groups by clicking on
the groups that you are a part of. I make the group home page as a easy
way to create and build the adding records functionality feature without
unnceessary hassle. This will also give me a way to see what is going on
with each of the groups that as the development process moves forward.
There was a bug when clicking on the group name links in the usres home so
I am manually entereing the groups name into the url. I will be fixing
this issue and the similar one I was having with the users home page. I
will be adding comments sooon for other peopel working on this project.