Commit: added records functionality
Explain:
I have added a working records function. this function allows a user to
create a new record for an expense within the group and add the amount and
description as well as secide whether the expense is an even split between
the users or individual expense between the users. This is a working
function but needs to be slightly modeified for optimal usage. The url
redirecting is still having an issue so i am still enterent the urls
manually. should be fixed once all of the sectionsa are running properly.
Comments will also be added soon.