Commit: Fixed Bugs with functionality
Explain:
There a few bugs when i was restructuring the add group and memebers
features to make the process alot more efficient. One of the bugs was the
adding the member foreignkey reference because there was no default value.
this was causing an un-nullable field and when i try to save it, django
throws an exception. It is fixed now and you can see everything that
changed. I also flushed the database to start with fresh data after the
model restructuring