Commit: Debugging and updating
Explain:
The two major changes that were made include updating and keeping track of
the group count as more members are added and debugging the adding records
functionality. So now when a new group member is added, the count for the
group table is updated and adds one to the current amount. There was an
issue with the group refernce with the records table that was occuring and
it was not allowing the user to add records, I have now updated the model
and tested it to make sure that the reference is set correctly and works
propertly. I am going to be adding all the comments for all of the
functionality in my next commit.