Commit: fixed a few bugs and commented code
Explain:
I fixed a but to properly display all of the groups information in the
group home page. Now it displays the members and records related to the
specific group.
I also commented the views file for other developers to more easily
understand what is going on in the views and follow what is going in
within the app.