Commit: Restructured the group model
Explain:
I decided to change the group model and switched the createdby column from
a integer that stored the users id that created the group to a foreign key
so that it stored the entire users object. I did this because it makes it
easier to reference different parts of the users info directly from within
the groups table object and not have ot make another queryset to find the
users informaiton. This way the entire object is stored within the groups
object.