Commit: Added new transaction function and style changes
Explain:
I Added a new transaction function to the records function. This new
function will be be the part of the record that actually saves and
calculates how much each person will need to pay for the expense that he
or she was added in. It will take the amount, description, and users
name. After the record is save with the correct amount, the amount will be
taken from the overall group expense and more importantly, the individuals
account balance through paypal.

I also make 2 small style changes to make the views more consistent. when
dealing with forms, the forms name will be stored in a variable called
form and passed to the html file as form to keep it all consistent