Commit: Added new transaction functionality
Explain:
I added a more complex functionality to the group record transactions. Now
when you go to create a record, it will display all of the group members
for the user to select all of the members of the group that are related to
the specific transaction. You will also have the ability to make the split
even or individual depending on how to you want to split. This will help
on the html file where you actually enter the amounts.