Commit: added even split funcitonality
Explain:
I have added the feature that will let the user basically add a record
with selected users. Once all the users are selected, the user can then
select the even split for the record. the user will then be redirected to
a pgae where he or she will enter the total amount due and a description.
Opent tab will then take the amount and number of members with this
transaction. The amount will then be split evently and the amount wil be
saved in the transactions records before it is verified and pushed to
PAYPAL so the user can get reimbursed for the amount he or she payed. The
transaction is also created and records individial amount so that paypal
will then be able to take the specified amount from each persons paypal
account later on.