Commit: added split function
Explain:
SO far i have added the even split function to the project in which the
user will be able to add a record and select the members of the group tha
twill be participating in the record before getting redirected to the
record page. in the record page, the user will then add the total amount
due and a description of the expense. once the record is submitted, the
bill will be split evenly but the number of members that there are in the
record and assign the amount to each users transaction amount.