Commit: updated entire applicaiton
Explain:
FOr this latest commit, I make modifications to the login and logout
system for the user in order to make it easier for the user and
devleopers. I also added the feature that makes sure that users are logged
in in order to access the application and if a user tries to access the
application without being logged in, it will redirect the user to the
login page. you can find the three lines of code at the beginning of each
of the views. Another change that was made is the utilizations of the
session to keep track of the user as he is logged in and keeps track of
him whereever he goes within the application. this makes it easier to
display his info no matter where he is, when he takes actions, have the
ability to look up any information that is realted to him in a fast and
simple way.