Commit: Added requests functionality
Explain:
I added the request funtionality to opentab. User can now search fo other
users and send them friend requests. It added and created a new record in
the requests table that will hold all of the requests for the app. In the
user home, you will be able to see all of the friend requests that you
have as well as the friend request that you have sent which are still
pending for the other user to accept or deny.