Commit: Full integrated the friends functionality into project
Explain:
After finishing the inital version of the relations portion of the
project, I have fully integrated this feature in to the rest of the
proejnct. The biggest change that occured is that whne you are adding
members to a group, it will display all of your friends and you check the
firends that you want in the group. Before, it was all of the members of
the application but not it is just friends. It also auto includes the
current user when new group members are added because that was not the
case when i first changed add memebers functionality.