Commit: Fixed small bug
Explain:
I was having a bug with the searching for users. if the user did not exist
and was searched, there was an error page that was showing up and throwing
excpetions/ I had to chagne the way the search was being handled/ instead
of just putting the entered name into the query set and search for the
matiching record which was causing an error with the query of the name did
not exist. With the new methods, the name that is saved as a variable and
is now testing against all the existing users. if there is a match, the
reocrd is sent to the html template, otherwise an error message is
displayed and passed.