Commit: added individual split transactions feature
Explain:
I got the individual split of an expenses up and running. The user can now
select indivudal split of the bill where he will then enter each person
amount, tax, and tip. THe view will then go though and assign the correct
values to the correct people. The tax and tip have not been encorporated
yet. I just wanted ot get the hard part working correctly. I also updated
a small bug that was occuring when the user was searching for a different
user. It is working now... The next thing I am going to do is the tip and
tax for the transactions...