Commit: added a basic prfofile function
Explain:
I added a profile function so that the user can add more informaiton about
himself or herself so that when a user is searches, the user can see more
information and mkae sure that it is the right person they are looking
for. Also I had to make a change hwere the privacy selection is selected.
It was cuasing an error so instead of having the options of privacy in te
models file, I changed it to the forms file and added the radio buttons
within as the form is generated which fixed the error.