Commit: Added Functionality
Explain:
Opentab not offers the ability for a user to keep track and manage the
money that is being trasnfered in and out of the app through paypal. you
will be able to see each transfer that is made through the app both for a
group and personal balances. There needs to be more work done to amk the
process more efficient but it is a working function. comments will be
added later...