Commit: completed balancing feature
Explain:
I have got a fully functioning version of the balancing  functionality for
the project. Now you can add a balance to your main account. the money
that is sent to your main account can also be sent to different groups
that will take money from the main account to the gorup accounts. Every
expense that is added to the group, it will then be deducted from the
group balance. I will add comments once I finish working on all of the
functionality details that need to be done...