Commit: Added validation to balances
Explain:
I added a portion of validation before money is transfered between a users
main account and a users group account. If the user tries to transfer
money from main account to group account and there are not enough funds,
it will throw an error. Same thing happens if the user tries to transfer
money from group account to main account and does not have neough funds.

Commenting is going to be the next main commit.