Commit: Restructured money transfer methods
Explain:
I have changed the way money is transfered and managed between a bank
account or linked account and the tabz account. There is no actual
integration until synapse is integreated into the project. But as of righ
tnow the user can transfer money(fake money, just a number value) between
imaginary bank account and imaginary tabz account. I have also manage to
integrate transfers into the activity or notificaiton part of the project
as well as a balance which keeps track of the transfers and expenses that
are created.... The next step is to integrate the transactions for each
person into the currently set up transfer and money balance functionality.