Commit: Added Synapse API first function
Explain:
I finally get the Synapse API connection to work and allow my self and the
project to create new users with the Synapse API. THis is done wiht two
parts, I added the header to the Client at the beginning of the views.py
file. The second part was creating a method that would pass user
informaitno and push it though the Synapse API and get a confirmation that
the user was infact created. This will ensire that the API is in fact
connecting with my project. Now it is all about storing the neccessary
information within the project database and addeing more functionality
wiht the Synpase API for more advanced development..