Commit: Creating user in synapse
Explain:
I have added a working version of creating and user and retreiving the information that is given back when the user is created in the synapse api. It also stores the info to be added to the database. I have also included the ability to input the uersn informaiton in the test api calls in order to make sure that each user account that is create is custom to each user.
`