Commit: coplete funtion integration - synapse user
Explain:
I have now full itegrated the creating new user funtion from Synapse API. opentab now has the ability to create a new user within the synapse api when a new profile is created. The api call also gives a response, and from within that response, an id is given that is specific to every synapse user. That id is then collected and stored in the local profile for ever user for later user lookup within the synapse api. I have also commented the new method that was added for other developers looking at this code.