Commit: Synapse linking bank account
Explain:
I have added the ability for the user to link a bank account through the synapse api service. The user will enter the bank, username, and password which will then be authenticated. It is currently commented out because there was an issue that kept occuring and was not allowing the authentication process to go thorughI am going to be adding the paypal api instead of the synpase api for the temporary testing periods.