Commit: Added new functionality - Dwolla accounts
Explain:
I have now implimented a means for the logged in user to link a bank account with their opentab and Dwolla accounts through the dwolla api. The user linked the bank account through the routing and account number. It is sent to the Dwolla api and the id from the response is stored locally. There is also the ability for the user to go into his linked accoutns and set the default account for fast processing. ALl of the code is commented for other developer to be able to understand what is going on