Commit: Added Dwolla Transactions
Explain:
Integrated the function of being able to submite transactions or transfers of money between two different users within the application. This is done by linking accounts, finding each users default account and then initiating the transfer. A trasnfer can not be created right now because the bank accounts are not verified within the sanbox testing phase.