Commit: Adding Synapse User Functions
Explain:
I have now added the ability for opentab to create a user within the Synapse api once a profile is submitted and completed. This will be usedful because it is required to create synapse accounts as well as transfers. The id for the user that is returns from the creting user method for synapse is saved in the users profile for easy reference later on within the apllication when needed: