Commit: Updated and did some restructuting
Explain:
THe biggest change that I made was how expenses were stored and created within the application. I also changed the database tables that store the information in toder to make it more manageable and scalable in the future. I changed a few specific things. Instead of a record and transaction table in the database and processing for both in the same method, I created one table and spereated the processing for easier processing and scalability. This is the big change that occuered. I also fixed the issue that was occuring with creating an individuallt split bill. it is not working.