Commit: Cleaned up the code
Explain:
In addition to the last commit, I fixed and setup the activity or notification portion fo the code so that it is working in a basic form. In this commit, I took out all of the code related to the record and transactions structure because it has been changed and made more efficient so it was a load of undeeded code that was clustering the file. THere is now no referecne to records and transactions anywhere.