Commit: Added new function
Explain:
I have added a new functiona that will allow you to create a transaction between two people in the application. It does not make the transation because within the sandbox, no transaction is authorized. The transaction will show up in the dashboard tho. I will be adding the transaction and connect it with the application.