Commit: Added viewing profiles and settings
Explain:
I have added the function so that when you are friends with someoen or search for someone, you can click on their username and view theri profile. it will display username, name, bio, and more. I started working on the settings page so that you can make changes to your user, password, info, and privacy settings. It still needs to be copleted which is going to be in the next commit. I will also be adding comments to the changed code once i have completed the functionality