Commit: added new functions
Explain:
FIrst and foremost, I reset the database for testing purposes and rebuilding the database and users. With functionality, the biggest new features is the validation of the expense through the users activity page. The user will now click on the actviities and where there is an expense, he or she will click the button on the side and that will verify the expense. all users have to verify expense before the money is transfered from account to account. I have also redone and restructured the activity objects that are created when new expenses are added, as well as when new members are added to the groups. Now when a member is added tot he gtoup, he will get the notification only. When an expense is added only the person with the expense gets a notificaiton and one goes to the group only. When the expense is verified, everyone gets a notification. I will be going in and commenting all of the code that I have added. I also worked on refineing the updating expenses methods for both the individual and even splits bills