Commit: Added functionality and css
Explain: I finally figured out how to add css to the files and start making change to the css properties for this project. I also manage to get the entire system running from activity to ente3ring infromation to changing group hosts. The enrire application should be accessable minus the Synapse nodes and transactions because I am still waying and trying to fix the problems that are coming up...