Commit: Began css and template integration
Explain:
I have restruture the template system to now encorporat a standard template system for the entire application that is applied. I have also began integrating custom css to the project for finer tooning. Ther is now a standard header and menu that will be used to direct and redirect the user to the correct pages.