Commit: Restructuredd template
Explain: I have now integrated a new base template for the application that is goig to be very close to the final version. It ameks it easier to work with the featues in teh near final form. I am completely changing the activity structures to make it more efficient and easier to work with from the backend.