Commit: Updated functionality
Explain:
I have worked on updting the activity section of the project and have begun getting it to the final stages of development and ready for final production. I have also make a few resturctures to the templates and how things are displayed and stored for the user depending on who is logged in at the moment. I am going to be working on the synapse integration and getting it to connect and work with Yap