Commit: Completed the Single Expense validation
Explain:
I have now completed the sungle expense validation processing. Now when a user creates a single expnse it will be processed the same way - forms. You will get a notificaition that there was a expense that was created and the amount you owe. when you confirm the expense you will get a new notificaion with the transfer when it is completed. .